Move from auto to sdk module
